refactor analysis scripts to reduce complexity
