+++ TIMES-NZ-INTERNAL-QA/data/concordances/commodities/commodity_sets_and_units.csv
@@ -91,66 +91,45 @@ COMWOD,NRG,PJ
 C_DATA-LT,DEM,PJ
 C_DATA-OTH,DEM,PJ
 C_DATA-SC,DEM,PJ
-C_EDU-Boiler-SH,DEM,PJ
-C_EDU-Boiler-WH,DEM,PJ
-C_EDU-Burner-SH,DEM,PJ
 C_EDU-LT,DEM,PJ
 C_EDU-MPM,DEM,PJ
 C_EDU-MPS,DEM,PJ
 C_EDU-OTH,DEM,PJ
 C_EDU-RF,DEM,PJ
-C_EDU-RH-SH,DEM,PJ
 C_EDU-SC,DEM,PJ
 C_EDU-SH,DEM,PJ
 C_EDU-WH,DEM,PJ
-C_HLTH-Boiler-PH,DEM,PJ
-C_HLTH-Boiler-SH,DEM,PJ
-C_HLTH-Boiler-WH,DEM,PJ
-C_HLTH-Burner-PH,DEM,PJ
 C_HLTH-LT,DEM,PJ
 C_HLTH-MPM,DEM,PJ
 C_HLTH-MPS,DEM,PJ
 C_HLTH-OTH,DEM,PJ
+C_HLTH-PH,DEM,PJ
 C_HLTH-RF,DEM,PJ
-C_HLTH-RH-PH,DEM,PJ
-C_HLTH-RH-SH,DEM,PJ
 C_HLTH-SC,DEM,PJ
 C_HLTH-SH,DEM,PJ
 C_HLTH-WH,DEM,PJ
-C_OFFC-Boiler-SH,DEM,PJ
-C_OFFC-Boiler-WH,DEM,PJ
-C_OFFC-Burner-SH,DEM,PJ
 C_OFFC-LT,DEM,PJ
 C_OFFC-MPM,DEM,PJ
 C_OFFC-MPS,DEM,PJ
 C_OFFC-OTH,DEM,PJ
-C_OFFC-RH-SH,DEM,PJ
 C_OFFC-SC,DEM,PJ
 C_OFFC-SH,DEM,PJ
 C_OFFC-WH,DEM,PJ
-C_OTH-Boiler-SH,DEM,PJ
-C_OTH-Boiler-WH,DEM,PJ
-C_OTH-Burner-PH,DEM,PJ
-C_OTH-Burner-SH,DEM,PJ
 C_OTH-LT,DEM,PJ
 C_OTH-MPM,DEM,PJ
 C_OTH-MPS,DEM,PJ
 C_OTH-OTH,DEM,PJ
+C_OTH-PH,DEM,PJ
 C_OTH-PUMP,DEM,PJ
-C_OTH-RH-SH,DEM,PJ
 C_OTH-SC,DEM,PJ
 C_OTH-SH,DEM,PJ
 C_OTH-WH,DEM,PJ
-C_WSR-Boiler-SH,DEM,PJ
-C_WSR-Boiler-WH,DEM,PJ
-C_WSR-Burner-SH,DEM,PJ
 C_WSR-CK,DEM,PJ
 C_WSR-LT,DEM,PJ
 C_WSR-MPM,DEM,PJ
 C_WSR-MPS,DEM,PJ
 C_WSR-OTH,DEM,PJ
 C_WSR-RF,DEM,PJ
-C_WSR-RH-SH,DEM,PJ
 C_WSR-SC,DEM,PJ
 C_WSR-SH,DEM,PJ
 C_WSR-WH,DEM,PJ
